reproductor/requirements.txt
Django
psycopg2
django-cors-headers
djangorestframework
Pillow
gunicorn